AstroBWTv3 support and better defaults settings:
* Cherry pick and squash commits from Tritonn:

  https://gitlab.com/Tritonn204/tnn-spectre-stratum-bridge/-/commit/79578a3d18ad7e58fb92ce8df86aa158f8265c5b (full)
  https://gitlab.com/Tritonn204/tnn-spectre-stratum-bridge/-/commit/79578a3d18ad7e58fb92ce8df86aa158f8265c5b (full)
  https://gitlab.com/Tritonn204/tnn-spectre-stratum-bridge/-/commit/2c31d50df138a7cf200e589b416b31ec87e3be55 (partial)

+++ src/spectrestratum/client_handler.go
@@ -146,7 +146,7 @@ func (c *clientListener) NewBlockAvailable(kapi *SpectreApi) {
 				jobParams = append(jobParams, GenerateLargeJobParams(header, uint64(template.Block.Header.Timestamp)))
 			} else {
 				jobParams = append(jobParams, GenerateJobHeader(header))
-				jobParams = append(jobParams, template.Block.Header.Timestamp)
+				jobParams = append(jobParams, uint64(template.Block.Header.Timestamp))
 			}
 
 			// // normal notify flow
